New Flight screens styles
We created all the screens for our app with the respective styles and
little funcionality, also we added navigation between the screes, the
first screen just have the blue arrow,  question,  input and button, the second has
the same estructur but we added the style for the info of the new flight, for the third screen we make it  with help from a library. the for the next screen we added a wheel-number for select he number of passanger and we add the styles for the blue arrow to the sides from the number selected. Finally  the user can be the data of their fligth and when you press the button, you will sent to the list of the flights.

Co-authored-by: JacquelineFM <jfloresm.dev@gmail.com>